REF: refactoring existing trace e2e tests
