+++ data-prepper-core/build.gradle
@@ -17,8 +17,6 @@ sourceSets {
     }
 }
 
-apply from: "integrationTest.gradle"
-
 dependencies {
     implementation project(':data-prepper-api')
     implementation project(':data-prepper-plugins')
