+++ e2e/trace/src/integrationTest/resources/data_prepper.yml
