+++ e2e/trace/src/integrationTest/resources/raw-span-e2e-pipeline-latest-release.yml
