+++ e2e/trace/src/integrationTest/resources/service-map-e2e-pipeline.yml
