+++ settings.gradle
@@ -37,4 +37,6 @@ include 'data-prepper-plugins:peer-forwarder'
 include 'data-prepper-plugins:blocking-buffer'
 include 'data-prepper-plugins:http-source'
 include 'data-prepper-plugins:grok-prepper'
+include 'e2e'
+include 'e2e:trace'
